clean up api a bit, goal: let the kernel do the font scaling
