+++ src/renderer_gl2/r_local.h
@@ -344,8 +344,8 @@ void R_DrawWorld (void);
 void R_RenderDlights(void);
 void R_DrawWorldAlphaSurfaces();
 void R_RenderBrushPoly (msurface_t *fa);
-void R_InitCodeTextures (void);
-void Draw_InitLocal (void);
+void R_InitCodeTextures();
+void R_LoadFonts();
 void GL_SubdivideSurface (msurface_t *fa);
 qboolean R_CullBox (vec3_t mins, vec3_t maxs);
 void R_RotateForEntity (rentity_t *e);
